edit cerner db config; implemented store sql and method

+++ src/main/scala/bh/DietNutrition.scala
@@ -1,6 +1,6 @@
 package bh
 
-case class DietNutrition(patientId: Int, encounterId: Int, status: String, diet: String) {
+case class DietNutrition(patientId: Long, encounterId: Long, status: String, diet: String) {
   def isValid: Boolean = patientId > 0 && encounterId > 0 && status.nonEmpty && diet.nonEmpty
 }
